# backend/app/routes/dashboards.py
"""
Dashboard routes for PowerBI Web Replica.
Handles dashboard creation, tile management, and layout operations.
"""

from fastapi import APIRouter, Depends, HTTPException, status
from sqlalchemy.ext.asyncio import AsyncSession
import structlog

from app.core.database import get_async_session
from app.core.security import get_current_user
from app.models.user import User

logger = structlog.get_logger()
router = APIRouter()


# Dashboard Routes (placeholder implementations)
@router.get("")
async def list_dashboards(
    workspace_id: str = None,
    current_user: User = Depends(get_current_user),
    session: AsyncSession = Depends(get_async_session)
):
    """List dashboards."""
    return {"dashboards": []}


@router.post("")
async def create_dashboard(
    current_user: User = Depends(get_current_user),
    session: AsyncSession = Depends(get_async_session)
):
    """Create new dashboard."""
    return {"message": "Dashboard creation not implemented yet"}


@router.get("/{dashboard_id}")
async def get_dashboard(
    dashboard_id: str,
    current_user: User = Depends(get_current_user),
    session: AsyncSession = Depends(get_async_session)
):
    """Get dashboard by ID."""
    return {"message": "Dashboard retrieval not implemented yet"}