# openapi.yml
openapi: 3.0.3
info:
  title: PowerBI Web Replica API
  description: |
    A comprehensive API for a web-based data visualization platform that recreates 
    Power BI Desktop functionality with select Tableau features.
    
    ## Features
    - Interactive report and dashboard design
    - Multiple data source connectors (CSV, PostgreSQL, etc.)
    - Drag-and-drop visualization builder
    - Real-time chart rendering with Vega-Lite
    - Offline mode support with IndexedDB
    - Export capabilities (PNG, PDF)
    - Role-based access control
    
  version: 1.0.0
  contact:
    name: PowerBI Web Replica
    url: https://github.com/powerbi-web-replica
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.powerbi-replica.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /api/workspaces/{workspace_id}/datasets:
    get:
      summary: List datasets in workspace
      tags: [Datasets]
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of datasets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dataset'

    post:
      summary: Create dataset (upload CSV or connect database)
      tags: [Datasets]
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: CSV file to upload
                name:
                  type: string
                  description: Dataset name
          application/json:
            schema:
              type: object
              required: [name, connector_type, connector_config]
              properties:
                name:
                  type: string
                  example: PostgreSQL Sales Data
                connector_type:
                  type: string
                  enum: [csv, postgresql, mysql, bigquery]
                  example: postgresql
                connector_config:
                  type: object
                  example:
                    host: localhost
                    port: 5432
                    database: sales
                    username: user
                    password: pass
      responses:
        '201':
          description: Dataset created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dataset'
  /api/datasets/{dataset_id}:
    get:
      summary: Get dataset details
      tags: [Datasets]
      parameters:
        - name: dataset_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Dataset details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dataset'
  /api/datasets/{dataset_id}/query:
    post:
      summary: Query dataset
      tags: [Datasets]
      parameters:
        - name: dataset_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                columns:
                  type: array
                  items:
                    type: string
                  example: ["category", "sales", "region"]
                filters:
                  type: array
                  items:
                    type: object
                    properties:
                      column:
                        type: string
                      operator:
                        type: string
                        enum: ["eq", "ne", "gt", "lt", "gte", "lte", "in", "like"]
                      value:
                        oneOf:
                          - type: string
                          - type: number
                          - type: array
                  example: [{"column": "region", "operator": "eq", "value": "North"}]
                aggregations:
                  type: array
                  items:
                    type: object
                    properties:
                      column:
                        type: string
                      function:
                        type: string
                        enum: ["sum", "count", "avg", "min", "max"]
                      alias:
                        type: string
                group_by:
                  type: array
                  items:
                    type: string
                order_by:
                  type: array
                  items:
                    type: object
                    properties:
                      column:
                        type: string
                      direction:
                        type: string
                        enum: ["asc", "desc"]
                limit:
                  type: integer
                  default: 1000
                offset:
                  type: integer
                  default: 0
      responses:
        '200':
          description: Query results
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  columns:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        type:
                          type: string
                  total_rows:
                    type: integer
                  execution_time:
                    type: number

  # Report Endpoints
...
  /api/dashboards:
    get:
      summary: List dashboards
      tags: [Dashboards]
      parameters:
        - name: workspace_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of dashboards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dashboard'

    post:
      summary: Create dashboard
      tags: [Dashboards]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workspace_id, name]
              properties:
                workspace_id:
                  type: string
                  format: uuid
                name:
                  type: string
                description:
                  type: string
                dashboard_json:
                  type: object
      responses:
        '201':
          description: Dashboard created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'

  # Export and Jobs
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Dataset:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        name:
          type: string
        connector_type:
          type: string
          enum: [csv, postgresql, mysql, bigquery]
        schema_json:
          type: object
          description: Inferred or configured data schema
        sample_rows:
          type: array
          description: Sample data for preview
        created_at:
          type: string
          format: date-time
        row_count:
          type: integer
        file_size:
          type: integer
...
    Dashboard:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        dashboard_json:
          type: object
          description: Dashboard layout and tile configuration
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...